separate collection interfaces and priorityqueue implementation
