+++ priorityqueue.go
@@ -2,83 +2,7 @@ package guc
 
 import (
 	"container/heap"
-	"time"
 )
-
-type Object interface {
-	Equals(i interface{}) bool
-}
-
-type Comparable interface {
-	CompareTo(i interface{}) int
-}
-
-type Comparator interface {
-	Compare(o1, o2 interface{}) int
-}
-
-type Iterator interface {
-	HasNext() bool
-	Next() interface{}
-	Remove()
-	ForEachRemaining(consumer func(i interface{}))
-}
-
-type Iterable interface {
-	Iterator() Iterator
-	ForEach(consumer func(i interface{}))
-}
-
-type Collection interface {
-	Iterable
-	Size() int
-	IsEmpty() bool
-	Contains(i interface{}) bool
-	ToArray() []interface{}
-	FillArray(arr []interface{}) []interface{}
-
-	Add(i interface{}) bool
-	Remove(i interface{}) bool
-	ContainsAll(coll Collection) bool
-	AddAll(coll Collection) bool
-	RemoveAll(coll Collection) bool
-	RemoveIf(predicate func(i interface{}) bool) bool
-	RetainAll(coll Collection) bool
-	Clear()
-	Equals(i interface{}) bool
-	HashCode() int
-}
-
-type Queue interface {
-	Collection
-
-	// default inherits
-	// Add(i interface{}) bool
-
-	Offer(i interface{}) bool
-	RemoveHead() interface{}
-	Poll() interface{}
-	Element() interface{}
-	Peek() interface{}
-}
-
-type BlockingQueue interface {
-	Queue
-
-	// default inherits
-	// Add(i interface{}) bool
-	// Offer(i interface{}) bool
-	// Remove(i interface{}) bool
-	// Contains(i interface{}) bool
-
-	Put(i interface{})
-	OfferWithTimeout(i interface{}, t time.Duration) bool
-	Take() interface{}
-	PollWithTimeout(t time.Duration) interface{}
-	RemainingCapacity() int
-	DrainTo(coll interface{}) int
-	DrainToWithLimit(coll interface{}, max int) int
-}
 
 var _ Queue = new(PriorityQueue)
